pyproject.toml
[project]
name = "bhashaai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "boto3>=1.42.34",
    "celery>=5.6.2",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.7",
    "langchain-cerebras>=0.8.2",
    "minio>=7.2.20",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]
